DataAnalytics by use of DAO with Hadoop MapReduce
Based on my white paper for reversing the customer Attrition by use of
Hadoop. this is the code repository for the same, It has implemented
with Pentahoo as GUI and processing was happened with DAO and MAP Reduce
was written for the Business purpose. After this, R code was written
separately in R studio, which is not in this code repository, Does the
Analytics for this Data Set Processed by MapReduce Job. Each time
Business will click the GUI, it invokes various Java as well as
MapRedcue Job